Audit fixes: 32-bit size-check regression, IGE KAT coverage
- tgcrypto.c: replace (Py_ssize_t) UINT32_MAX casts with uint64_t
  comparison in ige()/ctr()/cbc(). On 32-bit platforms the cast wraps
  to -1, making the 'Data is too large' check always true and breaking
  every API call. Verified via gcc -m32 assembly analysis.

- tests/test_safety.py: IGE known-answer test now accepts pycryptodome
  (Crypto) as well as pycryptodomex (Cryptodome), so the independent
  ECB-construction KAT no longer silently skips everywhere.

- wheels.yml: install pycryptodome in the CI test job so the IGE KAT
  runs in the test matrix.

Validation: 3886 passed / 0 skipped; AES-NI and software paths
bit-for-bit identical over 255 deterministic differential cases
(NIST vectors pass on both builds); outputs byte-identical pre/post
fix on all three modes.

+++ tgcrypto/tgcrypto.c
@@ -81,7 +81,9 @@ static PyObject *ige(PyObject *args, uint8_t encrypt) {
         return NULL;
     }
 
-    if (data.len > (Py_ssize_t) UINT32_MAX) {
+    /* Compare as uint64_t: casting UINT32_MAX to Py_ssize_t wraps to -1 on
+     * 32-bit platforms, which would reject every input. */
+    if ((uint64_t) data.len > (uint64_t) UINT32_MAX) {
         release3(&data, &key, &iv);
         PyErr_SetString(PyExc_ValueError, "Data is too large (max 4294967295 bytes)");
         return NULL;
@@ -162,7 +164,7 @@ static PyObject *ctr(PyObject *args) {
         return NULL;
     }
 
-    if (data.len > (Py_ssize_t) UINT32_MAX) {
+    if ((uint64_t) data.len > (uint64_t) UINT32_MAX) {
         release4(&data, &key, &iv, &state);
         PyErr_SetString(PyExc_ValueError, "Data is too large (max 4294967295 bytes)");
         return NULL;
@@ -226,7 +228,7 @@ static PyObject *cbc(PyObject *args, uint8_t encrypt) {
         return NULL;
     }
 
-    if (data.len > (Py_ssize_t) UINT32_MAX) {
+    if ((uint64_t) data.len > (uint64_t) UINT32_MAX) {
         release3(&data, &key, &iv);
         PyErr_SetString(PyExc_ValueError, "Data is too large (max 4294967295 bytes)");
         return NULL;
